refactor into miniapp
